+++ ZAPD/OtherStructs/CutsceneOoT_Commands.cpp
@@ -1,94 +1,11 @@
-#include "Cutscene_Commands.h"
-#include "Cutscene_CommandSubTypes.h"
+#include "CutsceneOoT_Commands.h"
+#include "CutsceneOoT_CommandSubTypes.h"
 
 #include <cassert>
 #include <unordered_map>
 #include "Utils/BitConverter.h"
 #include "Utils/StringHelper.h"
 
-
-/* CutsceneSubCommandEntry */
-
-CutsceneSubCommandEntry::CutsceneSubCommandEntry(const std::vector<uint8_t>& rawData,
-                                                 offset_t rawDataIndex)
-{
-	base = BitConverter::ToUInt16BE(rawData, rawDataIndex + 0);
-	startFrame = BitConverter::ToUInt16BE(rawData, rawDataIndex + 2);
-	endFrame = BitConverter::ToUInt16BE(rawData, rawDataIndex + 4);
-	pad = BitConverter::ToUInt16BE(rawData, rawDataIndex + 6);
-}
-
-std::string CutsceneSubCommandEntry::GetBodySourceCode() const
-{
-	return StringHelper::Sprintf("CMD_HH(0x%04X, 0x%04X), CMD_HH(0x%04X, 0x%04X)", base, startFrame,
-	                             endFrame, pad);
-}
-
-size_t CutsceneSubCommandEntry::GetRawSize() const
-{
-	return 0x08;
-}
-
-/* CutsceneCommand */
-
-CutsceneCommand::CutsceneCommand(const std::vector<uint8_t>& rawData, offset_t rawDataIndex)
-{
-	numEntries = BitConverter::ToUInt32BE(rawData, rawDataIndex + 0);
-}
-
-CutsceneCommand::~CutsceneCommand()
-{
-	for (auto& entry : entries)
-	{
-		delete entry;
-	}
-}
-
-std::string CutsceneCommand::GetCommandMacro() const
-{
-	return StringHelper::Sprintf("CMD_W(0x%08X), CMD_W(0x%08X)", commandID, numEntries);
-}
-
-std::string CutsceneCommand::GenerateSourceCode() const
-{
-	std::string result;
-
-	result += GetCommandMacro();
-	result += ",\n";
-
-	for (auto& entry : entries)
-	{
-		result += "        ";
-		result += entry->GetBodySourceCode();
-		result += ",\n";
-	}
-
-	return result;
-}
-
-size_t CutsceneCommand::GetCommandSize() const
-{
-	size_t size = 0;
-	if (entries.size() > 0)
-	{
-		size = entries.at(0)->GetRawSize() * entries.size();
-	}
-	else
-	{
-		size = 0x08 * numEntries;
-	}
-	return 0x08 + size;
-}
-
-void CutsceneCommand::SetCommandID(uint32_t nCommandID)
-{
-	commandID = nCommandID;
-
-	for (auto& entry : entries)
-	{
-		entry->commandID = commandID;
-	}
-}
 
 /*** GENERIC ****/
